Make sure flows kind of work
