+++ core-ui/src/components/Clusters/components/KubeconfigUpload.js
@@ -1,22 +1,21 @@
 import React from 'react';
-import createEncoder from 'json-url';
 import jsyaml from 'js-yaml';
 import LuigiClient from '@luigi-project/client';
-import { FileInput, DEFAULT_MODULES } from 'react-shared';
+import { FileInput } from 'react-shared';
 import { MessageStrip } from 'fundamental-react';
 import { KubeconfigTextArea } from './KubeconfigTextArea/KubeconfigTextArea';
 import { addCluster, readFile } from '../shared';
+import { decompress, getConfigFromParams } from './getConfigFromParams';
 
 export function KubeconfigUpload({ setKubeconfig, setShowingAuthForm }) {
   const [showError, setShowError] = React.useState(false);
-  const encoder = React.useRef(createEncoder('lzma'));
 
   const initParams = LuigiClient.getNodeParams().init;
 
   React.useEffect(() => {
     if (!initParams) return;
     async function setKubeconfigIfPresentInParams() {
-      const params = await encoder.current.decompress(initParams);
+      const params = await decompress(initParams);
       if (params.kubeconfig) {
         handleKubeconfigAdded(params.kubeconfig);
       }
@@ -24,16 +23,6 @@ export function KubeconfigUpload({ setKubeconfig, setShowingAuthForm }) {
     setKubeconfigIfPresentInParams();
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [initParams]);
-
-  const getConfigFromParams = async () => {
-    if (!initParams) return {};
-    const decoded = await encoder.current.decompress(initParams);
-    const clusterConfig = {
-      ...decoded?.config,
-      modules: { ...DEFAULT_MODULES, ...(decoded?.config?.modules || {}) },
-    };
-    return clusterConfig;
-  };
 
   async function onKubeconfigUploaded(file) {
     setShowError(false);
@@ -55,7 +44,7 @@ export function KubeconfigUpload({ setKubeconfig, setShowingAuthForm }) {
     const clientCA = user['client-certificate-data'];
     const clientKeyData = user['client-key-data'];
     if (token || (clientCA && clientKeyData)) {
-      const config = await getConfigFromParams();
+      const config = await getConfigFromParams(initParams);
       const params = {
         kubeconfig,
         config,
